feat: use CSS custom variables in pygments CSS
This commit reduces pygments-solarized complexity by using CSS custom
variables. This way it'll be easier to unify colors with highlight.js.

This commit also adds a utility that generates the mostly final
pygments-solarized stylesheet, which should reduce toil when wanting to
change how pygments highlighting works.

This commit is relevant to #4 and #6.