feat(transform_selection): use selected text
This commit refactors `transform_selection` to have the JS code directly
return rendered selection text with `Selection.toString` instead of
return the HTML.

This change greatly simplifies the code as we offload the responsibility
of parsing the selected HTML code to the selection API. It also makes
the code more robust.

This change should make implementing smart markdown detection (#87)
easier.